pyproject.toml
[project]
name = "nonebot-plugin-openai"
version = "0.1.0"
description = "Nonebot2 插件，同步Openai所有功能"
authors = [
    {name = "AkashiCoin", email = "i@loli.vet"},
]
license = {text = "MIT"}
dependencies = ["nonebot2>=2.0.0rc2", "nonebot-adapter-onebot>=2.1.5", "openai>=1.3.5"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/AkashiCoin/nonebot-plugin-openai"
Repository = "https://github.com/AkashiCoin/nonebot-plugin-openai"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"